Split MutationWriter into it's own module
